Extracted songIntersectionCalculation function
